+++ src/app/(main)/(menu)/activity/_components/active-button.tsx
@@ -1,5 +1,5 @@
-import { ButtonProps, buttonVariants } from '@/components/ui/button'
 import { cn } from '@/lib/utils'
+import { ButtonProps, buttonVariants } from '@/components/ui/button'
 
 type ActiveButtonProps = {
   isActive?: boolean
